chore: type contant -> constant
