+++ create-docs.ts
@@ -1,7 +1,7 @@
 import typedoc from 'typedoc'
 
 const app = await typedoc.Application.bootstrapWithPlugins({
-    entryPoints: ['./src/index.ts'],
+    entryPoints: ['./src/mod.ts'],
     name: 'yaksok.ts',
 })
