+++ runtest.ts
@@ -1,4 +1,4 @@
-import { yaksok } from './src/index.ts'
+import { yaksok } from './src/mod.ts'
 
 yaksok(
     `
