+++ src/bridge/quickjs/index.ts
@@ -9,7 +9,7 @@ import {
     type ValueTypes,
 } from '../../node/index.ts'
 import { bold, dim } from '../../error/common.ts'
-import type { FunctionParams } from '../../contant/type.ts'
+import type { FunctionParams } from '../../constant/type.ts'
 
 export class QuickJS {
     instance: QuickJSWASMModule | null = null
