+++ test/errors/exit-function-out-of-function.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { CannotReturnOutsideFunctionError } from '../../src/error/index.ts'
 
 Deno.test('약속의 밖에서는 `약속 그만`을 쓸 수 없음', () => {
