+++ test/errors/exit-loop-out-of-loop.test.ts
@@ -1,6 +1,6 @@
 import { assertIsError, unreachable } from 'assert'
 
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { BreakNotInLoopError } from '../../src/error/index.ts'
 
 Deno.test('반복의 밖에서는 `반복 그만`을 쓸 수 없음', () => {
