+++ test/errors/expected-argument-closing-parenthesis-but.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { UnexpectedTokenError } from '../../src/error/index.ts'
 
 Deno.test('인자를 닫는 괄호가 제 위치에 없습니다', () => {
