+++ test/errors/expected-function-body-but.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { UnexpectedTokenError } from '../../src/error/index.ts'
 
 Deno.test('온전하지 않은 약속', () => {
