+++ test/errors/formula.test.ts
@@ -1,5 +1,5 @@
 import { assertEquals } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 
 function createRandomValue(depth = 0): number | (string | number)[] {
     if (depth > 3 || Math.random() < 0.5) {
