+++ test/errors/function-without-static-parts.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { FunctionMustHaveOneOrMoreStringPartError } from '../../src/error/index.ts'
 
 Deno.test('고정된 부분이 없는 함수', () => {
