+++ test/errors/functions.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { InvalidTypeForOperatorError } from '../../src/error/index.ts'
 
 Deno.test('약속 안에서 발생한 오류', () => {
