+++ test/errors/invalid-character-for-code.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { UnexpectedCharError } from '../../src/error/index.ts'
 
 Deno.test('사용할 수 없는 문자', () => {
