+++ test/errors/invalid-parenthesis-content.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { CannotParseError } from '../../src/error/index.ts'
 
 Deno.test('올바르지 않은 괄호 묶음', () => {
