+++ test/errors/loop.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import {
     InvalidTypeForOperatorError,
     ListIndexOutOfRangeError,
