+++ test/errors/unexpected-end-of-parenthesis.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError, unreachable } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { UnexpectedEndOfCodeError } from '../../src/error/index.ts'
 
 Deno.test('끝나지 못한 괄호', () => {
