+++ test/errors/unexpected-eol.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { UnexpectedEndOfCodeError } from '../../src/error/index.ts'
 
 Deno.test('예상치 못한 줄바꿈', () => {
