+++ test/errors/unparsable-codes.test.ts
@@ -1,5 +1,5 @@
 import { assertIsError } from 'assert'
-import { yaksok } from '../../src/index.ts'
+import { yaksok } from '../../src/mod.ts'
 import { CannotParseError } from '../../src/error/index.ts'
 
 Deno.test('Unparsable codes', () => {
